Agregado graficacion 2D parte 1

+++ src/main/java/jeziel/graficadora/Modelos/Punto.java
@@ -1,11 +1,11 @@
 package jeziel.graficadora.Modelos;
 
 public abstract class Punto <P extends Punto<P,V>, V extends Vector<V>>{
-    protected float ordenadaX;
-    protected float ordenadaY;
-    protected float ordenadaZ;
+    protected double ordenadaX;
+    protected double ordenadaY;
+    protected double ordenadaZ;
 
-    public abstract float obtenerDistanciaPunto(P p);
+    public abstract double obtenerDistanciaPunto(P p);
     public abstract V vectorHastaPunto(P p);
     public abstract P trasladarPunto(V v);
 }
